requirements.txt
APScheduler==3.11.2
asgiref==3.11.1
attrs==26.1.0
boto3==1.42.73
botocore==1.42.73
cachetools==4.2.2
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.2.12
django-environ==0.13.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.3.1
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tornado==6.1
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3